4396 rename helper
