+++ web-client/src/presenter/state.js
@@ -50,7 +50,7 @@ import { orderTypesHelper } from './computeds/orderTypesHelper';
 import { pdfPreviewModalHelper } from './computeds/PDFPreviewModal/pdfPreviewModalHelper';
 import { pdfSignerHelper } from './computeds/pdfSignerHelper';
 import { requestAccessHelper } from './computeds/requestAccessHelper';
-import { reviewPetitionFromPaperHelper } from './computeds/reviewPetitionFromPaperHelperperHelper';
+import { reviewPetitionFromPaperHelper } from './computeds/reviewPetitionFromPaperHelper';
 import { scanBatchPreviewerHelper } from './computeds/scanBatchPreviewerHelper';
 import { scanHelper } from './computeds/scanHelper';
 import { selectDocumentSelectHelper } from './computeds/selectDocumentSelectHelper';
